간소화 및 Component 분리
