+++ 01-starting-project/src/components/CoreConcept.jsx
@@ -1,0 +1,11 @@
+function CoreConcept({ title, description, image }) {
+  return (
+    <li>
+      <img src={image} alt={title} />
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </li>
+  )
+}
+
+export { CoreConcept };
