+++ 01-starting-project/src/components/Header/Header.jsx
@@ -1,0 +1,17 @@
+import reactImg from '../../assets/react-core-concepts.png';
+import './Header.css'
+
+function Header() {
+  return (
+    <header>
+      <img src={reactImg} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        Fundamental React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
+
+export default Header;
